safari-534.53-branch/LayoutTests/sputnik/Conformance/07_Lexical_Conventions/7.3_Line_Terminators/S7.3_A6_T4-expected.txt

CONSOLE MESSAGE: line 76: SyntaxError: Parse error
S7.3_A6_T4

PASS Expected parsing failure

TEST COMPLETE
